bridge-adaptivity/requirements.txt
# install base requirements
-r requirements_base.txt

# additional deployment requirements
psycopg2
django-storages
boto

# temporary for debugging
# django-debug-toolbar